shouldHide for searchText set from non-user input.
The suggestions list should not show when setting the value for md-search-text from non-user input. 

As far as I understand, by checking to see if the input area has focus, we can determine that the change is being triggered by the end-user or not. Without this change, when binding a model from an async operation, the suggestions lists would trigger to show, resulting in the user seeing multiple suggestion lists showing when presented the form.

This proposed change also prevents an error when checking isMinLengthMet for a null value.